refactor(cli)!: restructure management flags into subcommands
- Move management flags into hierarchical noun subcommands: config
  init/path, auth login/logout, session list, skills path
- Keep agent main-flow flags on the default command (provider, model,
  turn guards, resume, json-stream, compaction, logging, etc.)
- Split the 690-line main.rs into focused modules: cli, run,
  bootstrap (shared config/logging/engine init), commands facade +
  per-group handlers, and a json_stream module tree
- Extract shared engine bootstrap to remove run/json_stream duplication;
  preserve the terminal-only "Resumed session" banner via a callback
- Update user-facing error strings and docs to reference the new
  subcommands instead of the removed flags

BREAKING CHANGE: the --init-config, --config-path, --login, --logout,
--list-sessions, and --skills-path flags are removed. Use the
corresponding subcommands (config init, config path, auth login, auth
logout, session list, skills path) instead.
