+++ crates/aion-config/src/auth.rs
@@ -197,7 +197,7 @@ impl OAuthManager {
             return Ok(new_creds.access_token);
         }
 
-        anyhow::bail!("Token expired and no refresh token available. Run 'aionrs --login'")
+        anyhow::bail!("Token expired and no refresh token available. Run 'aionrs auth login'")
     }
 
     /// Refresh the access token
@@ -254,7 +254,7 @@ impl OAuthManager {
 
     fn load_credentials(&self) -> anyhow::Result<OAuthCredentials> {
         let json = std::fs::read_to_string(&self.credentials_path)
-            .map_err(|_| anyhow::anyhow!("No saved credentials. Run 'aionrs --login'"))?;
+            .map_err(|_| anyhow::anyhow!("No saved credentials. Run 'aionrs auth login'"))?;
         let creds: OAuthCredentials = serde_json::from_str(&json)?;
         Ok(creds)
     }
